requirements.txt
# Vision & behavior dependencies

# Let pip also search the local Spot SDK wheel directory (sibling repo)
--find-links ../spot-sdk/prebuilt

# Boston Dynamics Spot SDK (installed from local wheels above)
bosdyn_api==5.0.1.2
bosdyn_client==5.0.1.2
bosdyn_core==5.0.1.2
bosdyn_mission==5.0.1.2
bosdyn_orbit==5.0.1.2
bosdyn_scout==5.0.1.2
bosdyn_choreography_client==5.0.1.2
bosdyn_choreography_protos==5.0.1.2

# Core
numpy
opencv-python
protobuf
Pillow

# YOLO person detection
ultralytics

# TensorFlow (for Network Compute worker)
tensorflow

# Note: TensorFlow Object Detection API (for label_map_util) must be installed separately.
